feat: move wasm files to separate dir
