+++ web/public/index.html
@@ -123,7 +123,7 @@
       </div>
     </div>
   </div>
-  <script src="/wasm_exec.js"></script>
+  <script src="/wasm/wasm_exec.js"></script>
   <% if (process.env.NODE_ENV === 'production') { %>
   {{ if .GoogleTagID }}
   <script async src="https://www.googletagmanager.com/gtag/js?id={{.GoogleTagID}}"></script>
