+++ web/src/service-worker.ts
@@ -70,7 +70,7 @@ registerRoute(
 );
 
 // Cache WebAssembly and Go assets
-const goWasmAssetsRegExp = new RegExp('^/(wasm_exec.js|go-repl.wasm|worker.wasm)$');
+const goWasmAssetsRegExp = new RegExp('^/wasm/(.*)(.js|.wasm)$', 'i');
 const DAY_IN_SECONDS = 24 * 60 * 60;
 registerRoute(
   ({ url }) => url.origin === self.location.origin && goWasmAssetsRegExp.test(url.pathname),
