+++ web/src/services/api/index.ts
@@ -2,5 +2,5 @@ export { default } from "./singleton";
 export * from "./models";
 export * from "./client";
 export * from "./interface";
-export * from "./utils";
+export * from "./resources";
 export * from "./provider";
