+++ web/src/services/api/resources.ts
@@ -1,6 +1,6 @@
 const {
   REACT_APP_WASM_API_VER: wasmApiVersion = 'v1',
-  REACT_APP_WASM_BASE_URL: wasmBaseUrl = '',
+  REACT_APP_WASM_BASE_URL: wasmBaseUrl = '/wasm',
 } = process.env;
 
 /**
@@ -9,3 +9,8 @@ const {
  * @param name
  */
 export const getWasmUrl = name => `${wasmBaseUrl}/${name}@${wasmApiVersion}.wasm`;
+
+/**
+ * URL for Go Wasm executor
+ */
+export const wasmExecUrl = `${wasmBaseUrl}/wasm_exec.js`;
